Cahcematrix Comments
The list created by makeCacheMatrix allows to acces the entire function's envirornment. This is key to being able to retreive the inverse.
Also by setting names to the list, we can acces the functions by name. This is very usefull in cacheSolve when it is necessary to set a new inverse